[pages/index][m]: refactor to use standard BaseSection component
